:recycle: #5 Refactor: 도메인(회원) Notification 테이블 Column 수정
