+++ src/main/java/com/umc/thegoods/domain/community/Inquiry.java
@@ -9,8 +9,6 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
-import java.util.ArrayList;
-import java.util.List;
 
 @Entity
 @Table(name = "inquiry")
@@ -31,8 +29,8 @@ public class Inquiry extends BaseDateTimeEntity {
     @JoinColumn(name = "item_id")
     private Item item;
 
-    @OneToMany(mappedBy = "inquiry", cascade = CascadeType.ALL)
-    private List<Inquiry> inquiryList = new ArrayList<>();
+    //@OneToMany(mappedBy = "inquiry", cascade = CascadeType.ALL)
+    //private List<Inquiry> inquiryList = new ArrayList<>();
 
     @Enumerated(EnumType.STRING)
     @Column(columnDefinition = "VARCHAR(10)", nullable = false)
