+++ src/main/java/com/umc/thegoods/domain/types/AlarmType.java
@@ -1,0 +1,5 @@
+package com.umc.thegoods.domain.types;
+
+public enum AlarmType {
+    MESSAGE, MENTION, ITEM, MARKETING, SELL_STATUS
+}
