--- src/main/java/com/umc/thegoods/mapping/member/MemberNotification.java
@@ -1,33 +1,0 @@
-package com.umc.thegoods.mapping.member;
-
-import com.umc.thegoods.domain.common.BaseDateTimeEntity;
-import com.umc.thegoods.domain.member.Member;
-import com.umc.thegoods.domain.member.Notification;
-import lombok.*;
-
-import javax.persistence.*;
-
-@Entity
-@Getter
-@Builder
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-public class MemberNotification extends BaseDateTimeEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "member_notification_id")
-    private Long id;
-
-    //알람 동의 여부
-    @Column(nullable = false, columnDefinition = "VARCHAR(1) DEFAULT 'X'")
-    private String agree;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id")
-    private Member member;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "notification_id")
-    private Notification notification;
-}
